autoregressive/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
simplejson==3.19.2
pandas==1.5.3
fastapi==0.104.0
uvicorn==0.23.2
Jinja2==3.1.2
rpy2==3.5.1